requirements.txt
# Моделът Gemini
google-genai
# Системата за сайта
streamlit